DockerImage/code/requirements.txt
beautifulsoup4
html2text
langid
roman
joblib
pathlib2
python-dotenv
